Fix twMerge for custom classnames
